+++ service/controller/resource/etcd-certificates/resource.go
@@ -1,8 +1,6 @@
 package etcdcertificates
 
 import (
-	"context"
-	"io/ioutil"
 	"reflect"
 
 	"github.com/giantswarm/k8sclient/v7/pkg/k8sclient"
@@ -12,7 +10,6 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	"github.com/giantswarm/prometheus-meta-operator/service/controller/resource/generic"
-	"github.com/giantswarm/prometheus-meta-operator/service/key"
 )
 
 const (
@@ -24,10 +21,11 @@ type Config struct {
 	Logger    micrologger.Logger
 }
 
-// secretCopier provides a `ToCR` method which copies data from the source
-// cluster secret CR
+// secretCopier provides a way to create a new secret from different data source.
 type secretCopier struct {
+	logger     micrologger.Logger
 	clientFunc func(string) generic.Interface
+	k8sClient  k8sclient.Interface
 }
 
 func New(config Config) (*generic.Resource, error) {
@@ -38,19 +36,24 @@ func New(config Config) (*generic.Resource, error) {
 		return nil, microerror.Maskf(invalidConfigError, "%T.Logger must not be empty", config)
 	}
 
+	// Wrapping the secret client into a generic interface.
 	clientFunc := func(namespace string) generic.Interface {
 		c := config.K8sClient.K8sClient().CoreV1().Secrets(namespace)
 		return wrappedClient{client: c}
 	}
 
-	sc := secretCopier{clientFunc: clientFunc}
+	sc := secretCopier{
+		logger:     config.Logger,
+		clientFunc: clientFunc,
+		k8sClient:  config.K8sClient,
+	}
 
 	c := generic.Config{
 		ClientFunc:       clientFunc,
 		Logger:           config.Logger,
 		Name:             Name,
 		GetObjectMeta:    getObjectMeta,
-		GetDesiredObject: sc.ToCR,
+		GetDesiredObject: sc.ToSecret,
 		HasChangedFunc:   hasChanged,
 	}
 	r, err := generic.New(c)
@@ -61,63 +64,7 @@ func New(config Config) (*generic.Resource, error) {
 	return r, nil
 }
 
-func getObjectMeta(v interface{}) (metav1.ObjectMeta, error) {
-	cluster, err := key.ToCluster(v)
-	if err != nil {
-		return metav1.ObjectMeta{}, microerror.Mask(err)
-	}
-
-	return metav1.ObjectMeta{
-		Name:      key.EtcdSecret(v),
-		Namespace: key.Namespace(cluster),
-	}, nil
-}
-
-func (sc *secretCopier) ToCR(v interface{}) (metav1.Object, error) {
-	objectMeta, err := getObjectMeta(v)
-	if err != nil {
-		return nil, microerror.Mask(err)
-	}
-
-	data, err := sc.getSource(context.TODO(), v)
-	if err != nil {
-		return nil, microerror.Mask(err)
-	}
-
-	secret := &corev1.Secret{
-		ObjectMeta: objectMeta,
-		StringData: data,
-	}
-
-	return secret, nil
-}
-
-// getSource returns the Secret to be copied, i.e. default/$CLUSTER_ID-prometheus
-func (sc *secretCopier) getSource(ctx context.Context, v interface{}) (map[string]string, error) {
-	ca, err := ioutil.ReadFile("/etcd-client-certs/ca.pem")
-	if err != nil {
-		return nil, microerror.Mask(err)
-	}
-
-	crt, err := ioutil.ReadFile("/etcd-client-certs/crt.pem")
-	if err != nil {
-		return nil, microerror.Mask(err)
-	}
-
-	key, err := ioutil.ReadFile("/etcd-client-certs/key.pem")
-	if err != nil {
-		return nil, microerror.Mask(err)
-	}
-
-	data := map[string]string{
-		"ca":  string(ca),
-		"crt": string(crt),
-		"key": string(key),
-	}
-
-	return data, nil
-}
-
+// hasChanged determines if secret data have changed.
 func hasChanged(current, desired metav1.Object) bool {
 	c := current.(*corev1.Secret)
 	d := desired.(*corev1.Secret)