+++ service/controller/resource/monitoring/ingress/v1/resource.go
@@ -1,6 +1,7 @@
 package v1
 
 import (
+	"context"
 	"fmt"
 	"reflect"
 
@@ -36,10 +37,10 @@ func New(config Config) (*generic.Resource, error) {
 		ClientFunc: clientFunc,
 		Logger:     config.Logger,
 		Name:       Name,
-		GetObjectMeta: func(v interface{}) (metav1.ObjectMeta, error) {
+		GetObjectMeta: func(ctx context.Context, v interface{}) (metav1.ObjectMeta, error) {
 			return getObjectMeta(v, config)
 		},
-		GetDesiredObject: func(v interface{}) (metav1.Object, error) {
+		GetDesiredObject: func(ctx context.Context, v interface{}) (metav1.Object, error) {
 			return toIngress(v, config)
 		},
 		HasChangedFunc: hasChanged,