+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add context to generic resources.
+
 ### Fixed
 
 - Add skip resource, to fix MC duplicated handling.